simplify data on tournament and move routes to ruby
