+++ app/javascript/RelayRouter/Routes.tsx
@@ -3,10 +3,7 @@ import React from 'react'
 import { graphql } from 'react-relay'
 import { MainLayout } from 'components/Layouts/MainLayout'
 import { App } from 'App'
-import { Home } from 'containers/Home'
 import { BracketList } from 'containers/BracketList'
-import { UserBracketList } from 'containers/UserBracketList'
-import { Bracket } from 'containers/Bracket'
 import { EditBracket } from 'containers/EditBracket'
 import { NewBracket } from 'containers/NewBracket'
 import { Games } from 'containers/Games'
@@ -17,22 +14,6 @@ const MainLayoutQuery = graphql`
   query Routes_MainLayout_Query {
     viewer {
       ...MainLayout_viewer
-    }
-  }
-`
-
-const UserBracketListQuery = graphql`
-  query Routes_UserBracketList_Query {
-    viewer {
-      ...UserBracketList_viewer
-    }
-  }
-`
-
-const BracketQuery = graphql`
-  query Routes_Bracket_Query($bracketId: ID!) {
-    bracket: node(id: $bracketId) {
-      ...Bracket_bracket
     }
   }
 `
@@ -55,19 +36,14 @@ const BracketListQuery = graphql`
 
 export default makeRouteConfig(
   <Route path="/" Component={App}>
-    <Route Component={Home} />
     <Route Component={MainLayout} query={MainLayoutQuery}>
       {/*    <Route path="possibilities" Component={Possibilities} query={PossibilitiesQuery} />*/}
       <Route path="brackets" Component={BracketList} query={BracketListQuery} />
-      <Route path="my_brackets" Component={UserBracketList} query={UserBracketListQuery} />
       <Route path="new_bracket" Component={NewBracket} />
       <Route path="games" Component={Games} />
       <Route path="rules" Component={RulesAndScoring} />
       <Route path="tournament/edit" Component={EditTournament} />
-      <Route path="brackets/:bracketId">
-        <Route Component={Bracket} query={BracketQuery} />
-        <Route path="edit" Component={EditBracket} query={EditBracketQuery} />
-      </Route>
+      <Route path="brackets/:bracketId/edit" Component={EditBracket} query={EditBracketQuery} />
     </Route>
   </Route>
 )
