+++ app/javascript/components/BasicBracket.ts
@@ -1,0 +1,7 @@
+export interface BasicBracket {
+  name: string
+  gameDecisions: bigint
+  gameMask: bigint
+}
+
+export const COMPLETED_MASK = 18446744073709551614n
