--- app/javascript/containers/UserBracketList/BracketStatus.tsx
@@ -1,8 +1,0 @@
-import React, { Component } from 'react'
-
-export const BracketStatus = ({ paid }: { paid: boolean }) => {
-  if (paid) {
-    return <span className="badge-success">OK</span>
-  }
-  return <span className="badge-alert">Unpaid</span>
-}
